requirements.txt
boto3==1.9.228
botocore==1.12.228
django-storages==1.7.2
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.0
s3transfer==0.2.1
urllib3==1.25.3
